add pwa aware router
