+++ src/SWP/Bundle/CoreBundle/Resources/config/services.yml
@@ -63,6 +63,12 @@ services:
         tags:
             - { name: router, priority: 300 }
 
+    SWP\Bundle\CoreBundle\Routing\PwaAwareRouter:
+        parent: "router.symfony"
+        public: true
+        tags:
+            - { name: router, priority: 300 }
+
     swp_core.provider.theme:
         class: SWP\Bundle\CoreBundle\Theme\Provider\ThemeProvider
         arguments:
